Only delete if ID already exists
FossilOrigin-Name: 8e4cb4776f8c00c7f23759bc2b5fabc0b09f80be759b397ef06984e89ae32111

+++ src/include/lib/Paheko/UserTemplate/Functions.php
@@ -356,11 +356,12 @@ class Functions
 		$document = $value;
 
 		if (!$result) {
-			// Always delete then replace
-			// in some rare cases (race condition) the item will already exist
-			// but will not be fetched in $result, in that case just overwrite what exists
 			$db->begin();
-			$db->delete($table, $field . ' = ?', $where_value);
+
+			if ($field && $where_value) {
+				$db->delete($table, $field . ' = ?', $where_value);
+			}
+
 			$db->insert($table, compact('id', 'document', 'key'));
 			$db->commit();
